+++ www/wp-content/plugins/jetpack/sync/class.jetpack-sync-modules.php
@@ -21,6 +21,7 @@ require_once dirname( __FILE__ ) . '/class.jetpack-sync-module-terms.php';
 require_once dirname( __FILE__ ) . '/class.jetpack-sync-module-plugins.php';
 require_once dirname( __FILE__ ) . '/class.jetpack-sync-module-protect.php';
 require_once dirname( __FILE__ ) . '/class.jetpack-sync-module-full-sync.php';
+require_once dirname( __FILE__ ) . '/class.jetpack-sync-module-stats.php';
 
 class Jetpack_Sync_Modules {
 
@@ -40,6 +41,7 @@ class Jetpack_Sync_Modules {
 		'Jetpack_Sync_Module_Plugins',
 		'Jetpack_Sync_Module_Protect',
 		'Jetpack_Sync_Module_Full_Sync',
+		'Jetpack_Sync_Module_Stats',
 	);
 
 	private static $initialized_modules = null;
@@ -78,13 +80,21 @@ class Jetpack_Sync_Modules {
 		 */
 		$modules = apply_filters( 'jetpack_sync_modules', self::$default_sync_modules );
 
-		return array_map( array( 'Jetpack_Sync_Modules', 'initialize_module' ), $modules );
+		$modules = array_map( array( 'Jetpack_Sync_Modules', 'load_module' ), $modules );
+
+		return array_map( array( 'Jetpack_Sync_Modules', 'set_module_defaults' ), $modules );
 	}
 
-	static function initialize_module( $module_name ) {
-		$module = new $module_name;
-		$module->set_defaults();
+	static function load_module( $module_name ) {
+		return new $module_name;
+	}
 
+	static function set_module_defaults( $module ) {
+		$module->set_defaults();
+		if ( method_exists( $module, 'set_late_default' ) ) {
+			add_action( 'init', array( $module, 'set_late_default' ), 90 );
+		}
 		return $module;
 	}
+
 }